plugins/speedtest/requirements.txt
speedtest-cli>=2.1.3
pillow>=8.0.0
